TEST/PERF: fix multi-thread performance
In the original, when multi-thread is enalbed, the performance is
worse than single thread.

In this commit, multiple ports are used to listen to the connection.
Firstly the port 13337 is used, when multiple thread is enabled,
the first thread connects to the port 13337.

Then the port 13347 is used to listen to the new connection. After
a new connection is comming. The port 13357 is used.

So the client can connect to the server with the different ports
concurrently.
This can increase the bandwidth totally.